Overhaul Windows drive inspection
