+++ lib/spack/spack/detection/common.py
@@ -21,6 +21,7 @@ import sys
 from typing import Dict, List, Optional, Set, Tuple, Union
 
 import llnl.util.tty
+from llnl.util.filesystem import _windows_drive
 
 import spack.config
 import spack.error
@@ -250,162 +251,6 @@ def set_virtuals_nonbuildable(virtuals: Set[str], scope: Optional[str] = None) -
     )
 
     return list(new_config.keys())
-
-
-def _windows_drive() -> str:
-    """Return Windows drive string extracted from the PROGRAMFILES environment variable,
-    which is guaranteed to be defined for all logins.
-    """
-    match = re.match(r"([a-zA-Z]:)", os.environ["PROGRAMFILES"])
-    if match is None:
-        raise RuntimeError("cannot read the PROGRAMFILES environment variable")
-    return match.group(1)
-
-
-class WindowsCompilerExternalPaths:
-    @staticmethod
-    def find_windows_compiler_root_paths() -> List[str]:
-        """Helper for Windows compiler installation root discovery
-
-        At the moment simply returns location of VS install paths from VSWhere
-        But should be extended to include more information as relevant"""
-        return list(winOs.WindowsOs().vs_install_paths)
-
-    @staticmethod
-    def find_windows_compiler_cmake_paths() -> List[str]:
-        """Semi hard-coded search path for cmake bundled with MSVC"""
-        return [
-            os.path.join(
-                path, "Common7", "IDE", "CommonExtensions", "Microsoft", "CMake", "CMake", "bin"
-            )
-            for path in WindowsCompilerExternalPaths.find_windows_compiler_root_paths()
-        ]
-
-    @staticmethod
-    def find_windows_compiler_ninja_paths() -> List[str]:
-        """Semi hard-coded search heuristic for locating ninja bundled with MSVC"""
-        return [
-            os.path.join(path, "Common7", "IDE", "CommonExtensions", "Microsoft", "CMake", "Ninja")
-            for path in WindowsCompilerExternalPaths.find_windows_compiler_root_paths()
-        ]
-
-    @staticmethod
-    def find_windows_compiler_bundled_packages() -> List[str]:
-        """Return all MSVC compiler bundled packages"""
-        return (
-            WindowsCompilerExternalPaths.find_windows_compiler_cmake_paths()
-            + WindowsCompilerExternalPaths.find_windows_compiler_ninja_paths()
-        )
-
-
-class WindowsKitExternalPaths:
-    @staticmethod
-    def find_windows_kit_roots() -> List[str]:
-        """Return Windows kit root, typically %programfiles%\\Windows Kits\\10|11\\"""
-        if sys.platform != "win32":
-            return []
-        program_files = os.environ["PROGRAMFILES(x86)"]
-        kit_base = os.path.join(program_files, "Windows Kits", "**")
-        return glob.glob(kit_base)
-
-    @staticmethod
-    def find_windows_kit_bin_paths(
-        kit_base: Union[Optional[str], Optional[list]] = None,
-    ) -> List[str]:
-        """Returns Windows kit bin directory per version"""
-        kit_base = WindowsKitExternalPaths.find_windows_kit_roots() if not kit_base else kit_base
-        assert kit_base, "Unexpectedly empty value for Windows kit base path"
-        if isinstance(kit_base, str):
-            kit_base = kit_base.split(";")
-        kit_paths = []
-        for kit in kit_base:
-            kit_bin = os.path.join(kit, "bin")
-            kit_paths.extend(glob.glob(os.path.join(kit_bin, "[0-9]*", "*\\")))
-        return kit_paths
-
-    @staticmethod
-    def find_windows_kit_lib_paths(
-        kit_base: Union[Optional[str], Optional[list]] = None,
-    ) -> List[str]:
-        """Returns Windows kit lib directory per version"""
-        kit_base = WindowsKitExternalPaths.find_windows_kit_roots() if not kit_base else kit_base
-        assert kit_base, "Unexpectedly empty value for Windows kit base path"
-        if isinstance(kit_base, str):
-            kit_base = kit_base.split(";")
-        kit_paths = []
-        for kit in kit_base:
-            kit_lib = os.path.join(kit, "Lib")
-            kit_paths.extend(glob.glob(os.path.join(kit_lib, "[0-9]*", "*", "*\\")))
-        return kit_paths
-
-    @staticmethod
-    def find_windows_driver_development_kit_paths() -> List[str]:
-        """Provides a list of all installation paths
-        for the WDK by version and architecture
-        """
-        wdk_content_root = os.getenv("WDKContentRoot")
-        return WindowsKitExternalPaths.find_windows_kit_lib_paths(wdk_content_root)
-
-    @staticmethod
-    def find_windows_kit_reg_installed_roots_paths() -> List[str]:
-        reg = spack.util.windows_registry.WindowsRegistryView(
-            "SOFTWARE\\Microsoft\\Windows Kits\\Installed Roots",
-            root_key=spack.util.windows_registry.HKEY.HKEY_LOCAL_MACHINE,
-        )
-        if not reg:
-            # couldn't find key, return empty list
-            return []
-        kit_root_reg = re.compile(r"KitsRoot[0-9]+")
-        root_paths = []
-        for kit_root in filter(kit_root_reg.match, reg.get_values().keys()):
-            root_paths.extend(
-                WindowsKitExternalPaths.find_windows_kit_lib_paths(reg.get_value(kit_root).value)
-            )
-        return root_paths
-
-    @staticmethod
-    def find_windows_kit_reg_sdk_paths() -> List[str]:
-        sdk_paths = []
-        sdk_regex = re.compile(r"v[0-9]+.[0-9]+")
-        windows_reg = spack.util.windows_registry.WindowsRegistryView(
-            "SOFTWARE\\WOW6432Node\\Microsoft\\Microsoft SDKs\\Windows",
-            root_key=spack.util.windows_registry.HKEY.HKEY_LOCAL_MACHINE,
-        )
-        for key in filter(sdk_regex.match, [x.name for x in windows_reg.get_subkeys()]):
-            reg = windows_reg.get_subkey(key)
-            sdk_paths.extend(
-                WindowsKitExternalPaths.find_windows_kit_lib_paths(
-                    reg.get_value("InstallationFolder").value
-                )
-            )
-        return sdk_paths
-
-
-def find_win32_additional_install_paths() -> List[str]:
-    """Not all programs on Windows live on the PATH
-    Return a list of other potential install locations.
-    """
-    drive_letter = _windows_drive()
-    windows_search_ext = []
-    cuda_re = r"CUDA_PATH[a-zA-Z1-9_]*"
-    # The list below should be expanded with other
-    # common Windows install locations as neccesary
-    path_ext_keys = ["I_MPI_ONEAPI_ROOT", "MSMPI_BIN", "MLAB_ROOT", "NUGET_PACKAGES"]
-    user = os.environ["USERPROFILE"]
-    add_path = lambda key: re.search(cuda_re, key) or key in path_ext_keys
-    windows_search_ext.extend([os.environ[key] for key in os.environ.keys() if add_path(key)])
-    # note windows paths are fine here as this method should only ever be invoked
-    # to interact with Windows
-    # Add search path for default Chocolatey (https://github.com/chocolatey/choco)
-    # install directory
-    windows_search_ext.append("%s\\ProgramData\\chocolatey\\bin" % drive_letter)
-    # Add search path for NuGet package manager default install location
-    windows_search_ext.append(os.path.join(user, ".nuget", "packages"))
-    windows_search_ext.extend(
-        spack.config.get("config:additional_external_search_paths", default=[])
-    )
-    windows_search_ext.extend(spack.util.environment.get_path("PATH"))
-    return windows_search_ext
 
 
 def compute_windows_program_path_for_package(pkg: "spack.package_base.PackageBase") -> List[str]:
